only annotate lookup from callset_ht
